+++ sentry-rails/lib/sentry/rails/capture_exceptions.rb
@@ -1,11 +1,11 @@
 module Sentry
   module Rails
     class CaptureExceptions < Sentry::Rack::CaptureExceptions
-      def initialize(app)
+      def initialize(_)
         super
 
-        if defined?(::Sprockets::Rails)
-          @assets_regex = %r(\A/{0,2}#{::Rails.application.config.assets.prefix})
+        if Sentry.initialized?
+          @assets_regexp = Sentry.configuration.rails.assets_regexp
         end
       end
 
@@ -25,13 +25,6 @@ module Sentry
 
         # the exception will be swallowed by ShowExceptions middleware
         return if request.show_exceptions? && !Sentry.configuration.rails.report_rescued_exceptions
-
-        current_scope = Sentry.get_current_scope
-
-        if original_transaction = env["sentry.original_transaction"]
-          current_scope.set_transaction_name(original_transaction)
-        end
-
         Sentry::Rails.capture_exception(exception).tap do |event|
           env[ERROR_EVENT_ID_KEY] = event.event_id if event
         end
@@ -39,13 +32,15 @@ module Sentry
 
       def start_transaction(env, scope)
         sentry_trace = env["HTTP_SENTRY_TRACE"]
-        options = { name: scope.transaction_name, op: transaction_op }
+        baggage = env["HTTP_BAGGAGE"]
 
-        if @assets_regex && scope.transaction_name.match?(@assets_regex)
+        options = { name: scope.transaction_name, source: scope.transaction_source, op: transaction_op }
+
+        if @assets_regexp && scope.transaction_name.match?(@assets_regexp)
           options.merge!(sampled: false)
         end
 
-        transaction = Sentry::Transaction.from_sentry_trace(sentry_trace, **options) if sentry_trace
+        transaction = Sentry::Transaction.from_sentry_trace(sentry_trace, baggage: baggage, **options) if sentry_trace
         Sentry.start_transaction(transaction: transaction, custom_sampling_context: { env: env }, **options)
       end
     end