+++ sentry-sidekiq/lib/sentry/sidekiq/sentry_context_middleware.rb
@@ -16,8 +16,8 @@ module Sentry
         scope.set_tags(queue: queue, jid: job["jid"])
         scope.set_tags(build_tags(job["tags"]))
         scope.set_contexts(sidekiq: job.merge("queue" => queue))
-        scope.set_transaction_name(context_filter.transaction_name)
-        transaction = start_transaction(scope.transaction_name, job["sentry_trace"])
+        scope.set_transaction_name(context_filter.transaction_name, source: :task)
+        transaction = start_transaction(scope, job["sentry_trace"])
         scope.set_span(transaction) if transaction
 
         begin
@@ -37,8 +37,8 @@ module Sentry
         Array(tags).each_with_object({}) { |name, tags_hash| tags_hash[:"sidekiq.#{name}"] = true }
       end
 
-      def start_transaction(transaction_name, sentry_trace)
-        options = { name: transaction_name, op: "sidekiq" }
+      def start_transaction(scope, sentry_trace)
+        options = { name: scope.transaction_name, source: scope.transaction_source, op: "sidekiq" }
         transaction = Sentry::Transaction.from_sentry_trace(sentry_trace, **options) if sentry_trace
         Sentry.start_transaction(transaction: transaction, **options)
       end